Add files via upload
The ever changing Hangman code. I found bugs in the last version, and eliminated them. I changed a lot of variable and function names to make it more readable and understandable to a reader. For the same reason, I reduced blocks of code to their own defined functions, to make the main body more readable.